requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
open_clip_torch>=2.20.0
hnswlib>=0.7.0
ultralytics>=8.0.0
streamlit>=1.25.0
Pillow>=9.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
